bnauth/pyproject.toml
[project]
name = "bnauth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.9",
    "ty>=0.0.29",
]
